Change 'sync-processed-sips' window 11 -> 31 days
When syncing processed SIPs using the DPRES SFTP endpoint, check for
SIPs from the last 31 days instead of just 11 days. The date range
doesn't really affect the runtime, so this is safe to increase and makes
the system more reliable if the synchronization is disabled for a longer
period of time for some reason.